+++ backend/crud/store_crud.py
@@ -1,0 +1,45 @@
+from models import User, Store, Location, Menu, Order
+from schemas import schemas
+from fastapi import Response
+from sqlalchemy.orm import Session
+from starlette.responses import Response
+from starlette.status import HTTP_201_CREATED, HTTP_204_NO_CONTENT
+
+
+# 가게 메뉴 정보 받기
+def get_store_menu(db: Session, store_id):  # 메뉴
+    menu = db.query(Menu).join(Store).filter(
+        Store.id == store_id).filter(Menu.is_active == True).all()
+    return menu
+
+
+# 가게 전체 조회
+def get_store(db: Session):  # 가게
+    return db.query(Store).all()
+
+
+# 가게 생성 (location table) 
+def create_store(db: Session, store: schemas.StoreCreate, loc: schemas.LocationCreate):
+    location = Location(points=loc.points)
+    db.add(location)
+    db.commit()
+    db_store = Store(user_id=store.user_id,
+                            location_id=location.id,
+                            name=store.name,
+                            category=store.category,
+                            description=store.description,
+                            photo_url=store.photo_url
+                            )
+
+    db.add(db_store)
+    db.commit()
+    return db_store
+
+
+# 가게 삭제
+def delete_store_by_id(db: Session, store_id: int):
+    store = db.query(Store).filter(Store.id ==
+                                          store_id).update({'is_active': False})
+    db.commit()
+    return Response(status_code=HTTP_204_NO_CONTENT)
+
